move explore page to explore package
