+++ packages/data-portal-commons/src/index.ts
@@ -1,6 +1,11 @@
 export * from './lib/entity';
 export * from './lib/entityReportHelpers';
+export * from './lib/fetchSynData';
+export * from './lib/fillInEntities';
+export * from './lib/filterHelpers';
 export * from './lib/getCaseValues';
+export * from './lib/isLowestLevel';
+export * from './lib/synapse';
 export * from './lib/types';
 
 export * from './components/ExpandableText';
