+++ packages/data-portal-commons/src/lib/isLowestLevel.ts
@@ -1,0 +1,5 @@
+import { BaseSerializableEntity } from './entity';
+
+export function isLowestLevel(entity: BaseSerializableEntity) {
+    return entity.Islowestlevel?.toLowerCase().startsWith('yes');
+}
