+++ packages/data-portal-commons/src/lib/synapse.ts
@@ -1,0 +1,16 @@
+export type SynapseData = {
+    atlases: SynapseAtlas[];
+};
+
+export type SynapseAtlas = {
+    htan_id: string;
+    htan_name: string;
+} & {
+    [data_schema: string]: SynapseRecords;
+};
+
+export type SynapseRecords = {
+    data_schema: string;
+    record_list: { values: any[] }[];
+    column_order: string[];
+};
