+++ packages/data-portal-explore/src/components/explore.module.scss
@@ -1,0 +1,3 @@
+.explore {
+    padding: 20px;
+}
