+++ packages/data-portal-explore/src/components/fileFilterControls.module.scss
@@ -1,0 +1,8 @@
+.filterCheckboxListContainer {
+    display: flex;
+
+    > div {
+        margin-right: 20px;
+        min-width: 300px;
+    }
+}
