+++ packages/data-portal-filter/src/components/FilterControls.tsx
@@ -4,6 +4,8 @@ import { getOptionsFromProps, getSelectOptionsFromProps } from '../lib/helpers';
 import { IGenericFilterControlProps } from '../lib/types';
 import FilterSearch from './FilterSearch';
 
+import styles from './filterControls.module.scss';
+
 export const FilterControls = observer(
     <T, Attribute extends string>(
         props: React.PropsWithChildren<IGenericFilterControlProps<T, Attribute>>
@@ -12,7 +14,7 @@ export const FilterControls = observer(
         const selectOptions = getSelectOptionsFromProps(props, options);
 
         return (
-            <div className="filterControls">
+            <div className={styles.filterControls}>
                 <FilterSearch
                     selectOptions={selectOptions}
                     setFilter={props.setFilter}
