+++ pages/tools.module.scss
@@ -1,0 +1,3 @@
+.tools {
+    padding: 20px;
+}
